Adding test for IllegalState being processed correctly
Issue: #529

Separated SW360UpdaterTest into two classes, one with
parameterized tests and one without. This allows for simple
tests on funcitonality that is not relying on the clearing
state.
